HW6: refactored jdbc user repository

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserRepositoryImpl.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.support.DataAccessUtils;
 import org.springframework.jdbc.core.BatchPreparedStatementSetter;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.ResultSetExtractor;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
@@ -24,8 +23,6 @@ import java.util.*;
 @Repository
 @Transactional(readOnly = true)
 public class JdbcUserRepositoryImpl implements UserRepository {
-
-    private static final BeanPropertyRowMapper<User> ROW_MAPPER = BeanPropertyRowMapper.newInstance(User.class);
 
     private final JdbcTemplate jdbcTemplate;
 
@@ -51,12 +48,15 @@ public class JdbcUserRepositoryImpl implements UserRepository {
         if (user.isNew()) {
             Number newKey = insertUser.executeAndReturnKey(parameterSource);
             user.setId(newKey.intValue());
-        } else if (namedParameterJdbcTemplate.update(
-                "UPDATE users SET name=:name, email=:email, password=:password, " +
-                        "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
-        }
+        } else {
+            if (namedParameterJdbcTemplate.update(
+                    "UPDATE users SET name=:name, email=:email, password=:password, " +
+                            "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
+                return null;
+            }
 
-        jdbcTemplate.update("DELETE FROM user_roles WHERE user_id=?", user.getId());
+            jdbcTemplate.update("DELETE FROM user_roles WHERE user_id=?", user.getId());
+        }
 
         String sql = "INSERT INTO user_roles (user_id, role) VALUES (?, ?)";
 
@@ -87,54 +87,22 @@ public class JdbcUserRepositoryImpl implements UserRepository {
 
     @Override
     public User get(int id) {
-        List<User> users = jdbcTemplate.query("SELECT * FROM users LEFT JOIN user_roles ON users.id = user_roles.user_id WHERE id=?", new RowMapperWithRoles(), id);
+        List<User> users = jdbcTemplate.query("SELECT * FROM users LEFT JOIN user_roles ON users.id = user_roles.user_id WHERE id=?", new RowMapper(), id);
         return DataAccessUtils.singleResult(users);
     }
 
     @Override
     public User getByEmail(String email) {
-        List<User> users = jdbcTemplate.query("SELECT * FROM users LEFT JOIN user_roles ON users.id = user_roles.user_id WHERE email=?", new RowMapperWithRoles(), email);
+        List<User> users = jdbcTemplate.query("SELECT * FROM users LEFT JOIN user_roles ON users.id = user_roles.user_id WHERE email=?", new RowMapper(), email);
         return DataAccessUtils.singleResult(users);
     }
 
     @Override
     public List<User> getAll() {
-        Map<Integer, User> users = jdbcTemplate.query("SELECT * FROM users ORDER BY name, email", rs -> {
-            Map<Integer, User> data = new LinkedHashMap<>();
-
-            while (rs.next()) {
-                int id = rs.getInt("id");
-
-                User user = new User();
-                user.setId(id);
-                user.setEmail(rs.getString("email"));
-                user.setName(rs.getString("name"));
-                user.setCaloriesPerDay(rs.getInt("calories_per_day"));
-                user.setEnabled(rs.getBoolean("enabled"));
-                user.setPassword(rs.getString("password"));
-                user.setRoles(new LinkedHashSet<>());
-                data.putIfAbsent(id, user);
-            }
-
-            return data;
-        });
-
-        jdbcTemplate.query("SELECT * FROM user_roles", rs -> {
-            do {
-                User user = users.get(rs.getInt("user_id"));
-
-                if (user == null) {
-                    continue;
-                }
-
-                user.getRoles().add(Role.valueOf(rs.getString("role")));
-            } while (rs.next());
-        });
-
-        return new ArrayList<>(users.values());
+        return jdbcTemplate.query("SELECT * FROM users LEFT JOIN user_roles ON users.id = user_roles.user_id ORDER BY name, email", new RowMapper());
     }
 
-    public class RowMapperWithRoles implements ResultSetExtractor<List<User>> {
+    public class RowMapper implements ResultSetExtractor<List<User>> {
         @Override
         public List<User> extractData(ResultSet rs) throws SQLException, DataAccessException {
             Map<Integer, User> users = new LinkedHashMap<>();
@@ -154,7 +122,7 @@ public class JdbcUserRepositoryImpl implements UserRepository {
                     newUser.setCaloriesPerDay(calories);
                     newUser.setEnabled(enabled);
                     newUser.setPassword(password);
-                    newUser.setRoles(new LinkedHashSet<>());
+                    newUser.setRoles(EnumSet.noneOf(Role.class));
                     return newUser;
                 });
 
@@ -164,5 +132,4 @@ public class JdbcUserRepositoryImpl implements UserRepository {
             return new ArrayList<>(users.values());
         }
     }
-
 }
